为了在私有项目中支持微信小程序开发,在windows下面增加了支持https和websocket

+++ projects/libcrypt/src/ssl/base64.cpp
@@ -113,7 +113,7 @@ int base64_encode_EX( unsigned char *dst, int *dlen,
         *p++ = '=';
     }
 
-    *dlen = p - dst;
+    *dlen = (int)(p - dst);
     *p = 0;
 
     return( 0 );
@@ -178,7 +178,7 @@ int base64_decode_EX( unsigned char *dst, int *dlen,
         }
     }
 
-    *dlen = p - dst;
+    *dlen = (int)(p - dst);
 
     return( 0 );
 }
