+++ src/core/string/utf8tool.cpp
@@ -37,10 +37,10 @@ void Utf8Tool::UNICODE_to_UTF8(const CString& unicodeString, std::string& str)
 
 string Utf8Tool::UNICODE_to_UTF8(const CString& unicodeString)
 {
-	auto stringLength = ::WideCharToMultiByte(CP_UTF8, NULL, unicodeString, wcslen(unicodeString), NULL, 0, NULL, NULL);
+	auto stringLength = ::WideCharToMultiByte(CP_UTF8, NULL, unicodeString, (int)wcslen(unicodeString), NULL, 0, NULL, NULL);
 
 	char* buffer = new char[stringLength + 1];
-	::WideCharToMultiByte(CP_UTF8, NULL, unicodeString, wcslen(unicodeString), buffer, stringLength, NULL, NULL);
+	::WideCharToMultiByte(CP_UTF8, NULL, unicodeString, (int)wcslen(unicodeString), buffer, stringLength, NULL, NULL);
 	buffer[stringLength] = '\0';
 
 	string str = buffer;
